+++ adiscord/src/gateway/mod.rs
@@ -38,7 +38,7 @@ enum Call {
 struct GatewayClient {
     token: String,
     handle: ezsockets::Client<Self>,
-    intents: u16,
+    intents: u32,
     callbacks: HashMap<String, Callback>,
     heartbeat_ack: bool,
     heartbeat_count: u32,
